Upgrade from Jetty 9.0.x to 9.4.x
