+++ pom.xml
@@ -12,14 +12,27 @@
     <version>1.0.3-SNAPSHOT</version>
     <packaging>jar</packaging>
 
-    <name>dropwizard-couchbase-sessions</name>
+    <name>dropwizard-couchbase-sessions-dw1x</name>
     <description>HttpSession management for dropwizard applications using Couchbase.</description>
     <url>https://github.com/cvent/dropwizard-couchbase-sessions</url>
 
     <properties>
+        <dropwizard.version>1.1.3</dropwizard.version>
         <jackson.version>2.3.3</jackson.version>
         <jersey.version>1.18.1</jersey.version>
     </properties>
+
+    <dependencyManagement>
+        <dependencies>
+            <dependency>
+                <groupId>io.dropwizard</groupId>
+                <artifactId>dropwizard-bom</artifactId>
+                <version>${dropwizard.version}</version>
+                <type>pom</type>
+                <scope>import</scope>
+            </dependency>
+        </dependencies>
+    </dependencyManagement>
 
     <dependencies>
         <dependency>
@@ -30,12 +43,12 @@
         <dependency>
             <groupId>com.fasterxml.jackson.core</groupId>
             <artifactId>jackson-core</artifactId>
-            <version>${jackson.version}</version>
+            <!--version>${jackson.version}</version-->
         </dependency>
         <dependency>
             <groupId>com.fasterxml.jackson.core</groupId>
             <artifactId>jackson-databind</artifactId>
-            <version>${jackson.version}</version>
+            <!--version>${jackson.version}</version-->
         </dependency>
         <dependency>
             <groupId>org.eclipse.jetty.orbit</groupId>
@@ -45,12 +58,12 @@
         <dependency>
             <groupId>org.eclipse.jetty</groupId>
             <artifactId>jetty-server</artifactId>
-            <version>9.0.7.v20131107</version>
+            <!--version>9.0.7.v20131107</version-->
         </dependency>
         <dependency>
             <groupId>org.slf4j</groupId>
             <artifactId>slf4j-api</artifactId>
-            <version>1.7.6</version>
+            <!--version>1.7.6</version-->
         </dependency>
         <dependency>
             <groupId>com.sun.jersey</groupId>
